compiler warnings cosmetics

+++ src/tests/valid/data.c
@@ -113,7 +113,7 @@ static inline lzws_result_t test_eof_after_completed_bits_group(
 {
   lzws_code_fast_t first_code    = lzws_get_first_free_code(compressor_state_ptr->options.block_mode);
   lzws_code_fast_t target_number = (1 << LZWS_LOWEST_MAX_CODE_BIT_LENGTH) - first_code + 1;
-  size_t           string_length = target_number * (target_number + 1) / 2 + 2; // + 1 byte for \0.
+  size_t           string_length = (size_t)(target_number * (target_number + 1) / 2 + 2); // + 1 byte for \0.
   char             symbol        = 'a';
 
   char* string = lzws_allocate_array(1, string_length, (void*) &symbol, false, true);
